refactor: move stream test utilities to shared _test-utils package
