+++ server-adapters/_test-utils/src/index.ts
@@ -6,4 +6,4 @@ export {
   type RouteAdapterTestSuiteConfig,
 } from './route-adapter-test-suite';
 
-export { createDefaultTestContext } from './test-helpers';
+export { createDefaultTestContext, createStreamWithSensitiveData, consumeSSEStream } from './test-helpers';
